Rename compile-cache modules to consistent compile-cache-* prefix
  app/src/compile-cache.js          -> app/src/compile-cache-ts.js
  app/src/uncompiled-source-error.js -> app/src/compile-cache-ts-unsupported.js
  app/src/less-compile-cache.ts     -> app/src/compile-cache-less.ts

The runtime cache directory names on disk ('compile-cache',
'less-compile-cache') are left unchanged so existing user caches keep
working.
